+++ NEWS.md
@@ -3,6 +3,18 @@ journal-gateway-gelf
 
 A gateway for transmitting of systemds journal via to a graylog server.
 
+Version 1.0.2 - 02 Oct 2015
+---------------------------
+
+* fix libcurl usage
+* renaming gateway-gelf-source -> gateway-gelf
+
+Version 1.0.1 - 02 Oct 2015
+---------------------------
+
+* fix default config
+
+
 Version 1.0.0 - 01 Oct 2015
 ---------------------------